+++ server/src/main/java/com/uj/projects/booksplatform/controller/ExampleController.java
