+++ server/src/main/java/com/uj/projects/booksplatform/database/SQLiteDialect.java
