+++ server/src/main/java/com/uj/projects/booksplatform/repository/BookRepository.java
