+++ server/src/test/java/com/uj/projects/booksplatform/BooksplatformApplicationTests.java
